requirements.txt
torch>=2.0
transformers>=4.30
numpy
llama-cpp-python
